move slice utils from cli to rdk, add FilterSlice
